Added conditionals to Main to check Age, Weight and Height and print each object.

+++ app/src/main/java/com/example/kotlinfoundations/Individualable.kt
@@ -1,11 +1,8 @@
 package com.example.kotlinfoundations
 
-import kotlin.math.pow
-import kotlin.random.Random
-
 interface Individualable {
 
-    fun calculateBMI(weight:Int, height:Double): Int{
+    fun calculateBMI(weight: Double, height:Double): Int{
         val bmi: Double = weight.toDouble() / (height * height)
         return if (bmi < 18.5) {
             -1
